refactor: 타입 이름 수정
