+++ packages/client/src/pages/home/components/CoinView.tsx
@@ -2,7 +2,7 @@ import { useMarketAll } from '@/hooks/useMarketAll';
 import CoinCategories from '@/pages/home/components/CoinCategories';
 import CoinList from '@/pages/home/components/CoinList';
 import { MarketData } from '@/types/market';
-import { MarketCategory } from '@/types/menu';
+import { MarketCategory } from '@/types/category';
 import { filterCoin } from '@/utility/marketDataUtil';
 import { isMarket } from '@/utility/typeGuard';
 import { useState } from 'react';
