+++ packages/client/src/types/category.ts
@@ -1,5 +1,5 @@
 import { Market } from '@/types/market';
 
-export type SideBarMenu = 'MY_INVESTMENT' | 'INTEREST' | 'RECENTLY_VIEWED' | 'REALTIME' | null;
+export type SideBarCategory = 'MY_INVESTMENT' | 'INTEREST' | 'RECENTLY_VIEWED' | 'REALTIME' | null;
 
 export type MarketCategory = Market | 'INTEREST' | 'OWN';
